requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alembic==1.14.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==24.2.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
distlib==0.3.9
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
Mako==1.3.8
MarkupSafe==2.1.5
marshmallow==3.22.0
misaka==2.1.1
multidict==6.1.0
nba_api==1.5.2
numpy==1.24.4
openai==0.28.0
packaging==24.2
pipenv==2024.4.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.3.3
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
SQLAlchemy-serializer==1.4.12
tomli==2.1.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.2.3
virtualenv==20.27.1
Werkzeug==3.0.6
wsproto==1.2.0
yarl==1.15.2
yup==0.1.0
zipp==3.20.2
